+++ src/app/app.module.ts
@@ -5,9 +5,10 @@ import { AppComponent } from "./app.component";
 import { ProductsModule } from "./products/index";
 import { CartModule } from "./cart/index";
 import { SharedModule } from "./shared/index";
+import { AboutComponent } from "./about/index";
 
 @NgModule({
-  declarations: [AppComponent],
+  declarations: [AppComponent, AboutComponent],
   imports: [AppRoutingModule, CartModule, ProductsModule, SharedModule],
   bootstrap: [AppComponent],
 })