+++ src/app/core/models/config-options.model.ts
@@ -1,0 +1,8 @@
+/**
+ * Interface for config options
+ */
+export interface ConfigOptions {
+  id: number;
+  login: string;
+  email?: string;
+}