+++ src/app/core/services/config-options.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from "@angular/core/testing";
+
+import { ConfigOptionsService } from "./config-options.service";
+
+describe("ConfigOptionsService", () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it("should be created", () => {
+    const service: ConfigOptionsService = TestBed.get(ConfigOptionsService);
+    expect(service).toBeTruthy();
+  });
+});