+++ src/app/core/services/generator-factory.ts
@@ -1,0 +1,8 @@
+import { InjectionToken } from "@angular/core";
+import { GeneratorService } from "./generator.service";
+
+export const Generator = new InjectionToken<string>("Generator");
+
+export function GeneratorFactory(length: number) {
+  return (data: GeneratorService): string => data.generateData(length);
+}